chore: 🔄 update dependencies and enhance project configuration
This commit updates several dependencies in the project, including upgrading `expo` to version `54.0.30`, `expo-linking` to `8.0.11`, and `expo-router` to `6.0.21`. Additionally, it introduces the `motion` library for improved animations. The `app.json` file has been modified to include `expo-router` as a plugin, enhancing routing capabilities. These changes aim to ensure compatibility with the latest features and improve the overall functionality of the application.